requirements.txt
pandas==2.2.2
pymysql==1.1.1
tqdm==4.66.4
python-dotenv==1.0.1
cryptography==42.0.8